requirements.txt
feedparser
pyyaml
pymongo
